Start refactoring of displayProject
